New UI - panels
As time flies, wide screen are the norm. Lets abuse them:

The new UI is split in half, in the left side you see the OCRed text,
and on the right the "source". In the future, I will be able to load
"pfd" (or whatever) and I will be able to paginate between the pages and
OCR "on the fly".

The PixmapViewer is now defined in the UI file, which simplified code a
little.